Fix code scanning alert - Comparison of narrow type with wide type in loop condition #18

+++ src/bootloader/stage2/stdio.c
@@ -298,7 +298,7 @@ void print_buffer(const char* msg, const void* buffer, uint32_t count) {
 	const uint8_t* u8Buffer = (const uint8_t*)buffer;
 
 	puts(msg);
-	for (uint16_t i = 0; i < count; i++) {
+	for (uint32_t i = 0; i < count; i++) {
 		putc(g_HexChars[u8Buffer[i] >> 4]);
 		putc(g_HexChars[u8Buffer[i] & 0xF]);
 	}
